Make VS Code launch config a bit more useful
This commit is only interesting for people that use VS Code.

It adds a task that executes a makefile target. It changes the launch
config so that above task is launched before starting/attaching the
debugger.

The makefile sets all the necessary build flags. To avoid to specify
them again or having build flags divert across development environments
we teach VS Code just to use the makefile.

References:
- https://code.visualstudio.com/docs/editor/tasks
- https://code.visualstudio.com/docs/languages/go