remove `cobra.OnInitialize`
This func will add other func into a initialization queue. However, those funcs will not be executed in the `init` function, they will be invoked after the `init` func of the cmd package. This caused the nil pointer error since the `client` in `initStorageCmd` is nil (as it is not set yet). So that we have to create a client first in the main `init` func of the package.