Re-work error handling
Do not use os.Exit directly (or indirectly through log.Fatal or other
helper functions) anymore.

Motivation: we need to clean up resources in error cases (see
https://github.com/gridscale/gscloud/issues/97). os.Exit makes the
program terminate immediately and deferred functions are not run. This
makes it impossible to use defer for cleaning up real resources when we
need to bail out.

This commit replaces all Run functions with RunE, returning errors from
now on in subcommands instead of exiting the process immediately. Other
functions use panic for errors that shouldn't actually happen.

This commit also introduces cmd.Error as a custom error type. It
contains a What message that should explain to the user what went wrong
and an Err error with all the details (probably from the API).