Add RELEASE.txt file for archives
Use export-subst and git log's custom formatter to generate a file that
is included in every source tarball (and in every future release). From
this file, version information can be pulled and included included as
external build parameters.

Note that git-log and git-archive only learned about "%(describe)" in
not yet released 2.32.0.