Makefile: source RELEASE.txt when git fails
With commits 31f64b8 and 714dfbe and recently released Git v2.32 we
create a file RELEASE.txt that contains GIT_COMMIT and VERSION
variables. This should happen automatically when GitHub creates the
release archive or git-archive(1) is invoked directly.

To test, create an archive

    $ git archive --format=tar HEAD | gzip >gscloud.tar.gz

extract it and run make. The version strings produced by

    $ gscloud version

should match those for the original repository.

Fixes #131.