Document --config flag explicitly in README
When using the CLI, i found myself missing an equivalent to how tools like `kubectl` can be configured to
use a different config.

When searching for a similar option, i found the `--config` flag.

With this commit I want to explicitly document that flag
so that other users may find this option too.